system-tests: Extract C++ `trace` into separate file
Makes `stub.cpp` smaller and `trace` can be easier reused for debugging
purposes.
