+++ stub/cbits/trace.cpp
@@ -1,0 +1,16 @@
+#include <stdarg.h>
+#include <stdio.h>
+
+#ifdef TRACE
+void trace(const char *fmt, ...) {
+    va_list args;
+    va_start(args, fmt);
+    vprintf(fmt, args);
+    va_end(args);
+}
+#else
+void trace(const char * fmt, ...){
+  (void) fmt;
+}
+#endif
+
