+++ stub/cbits/trace.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void trace(const char *fmt, ...);
