+++ stub/ghc-debug-stub.cabal
@@ -23,7 +23,7 @@ library
   hs-source-dirs:      src
   build-depends:       base >=4.12 && <4.14, ghc-prim
   default-language:    Haskell2010
-  cxx-sources:         cbits/stub.cpp, cbits/socket.cpp
+  cxx-sources:         cbits/stub.cpp, cbits/socket.cpp, cbits/trace.cpp
   cxx-options:         -std=gnu++11
   ghc-options:         -threaded
   extra-libraries:     stdc++
